Struct oak_stage0::allocator::BumpAllocator

pub struct BumpAllocator<const N: usize> { /* private fields */ }

Basic bump allocator that never deallocates memory.

The algorithm is rather simple: we maintain an index into the buffer we hold, that marks the high water mark (already allocated memory). Whenever a request for memory comes in, me move the index forward by the padding necessary to maintain proper alignment for the structure plus the size of the structure.

We do not support deallocation, and there is no optimizations to pack the allocations as efficiently as possible.

For stage0 uses these limitations are not an issue because (a) we use this allocator to allocate data structures we expect to outlive stage0 and (b) we allocate only a small number of data structures from it, so the padding overhead is minimal.

Implementations

impl<const N: usize> BumpAllocator<N>

pub const fn uninit() -> Self