Trait oak_stage0::hal::Mmio

pub trait Mmio<S: PageSize> {
    // Required methods
    fn read_u32(&self, offset: usize) -> u32;
    unsafe fn write_u32(&mut self, offset: usize, value: u32);
}

Abstraction around MMIO (memory-mapped I/O) read/write access.

Normally you can just access the memory directly via read_volatile/write_volatile, but for SEV-ES and above we need to go via the GHCB to do MMIO.

Required Methods

fn read_u32(&self, offset: usize) -> u32

Reads an u32 from the MMIO memory region.

The offset is the number of u32-s (not byte offsets); that is, to read bytes at [base_address+4, base_address+8) the offset needs to be 1.

Panics if the read would go outside the memory range.

unsafe fn write_u32(&mut self, offset: usize, value: u32)

Write an u32 from the MMIO memory region.

The offset is the number of u32-s (not byte offsets); that is, to read bytes at [base_address+4, base_address+8) the offset needs to be 1.

Panics if the read would go outside the memory range.

The caller needs to guarantee that the value is valid for the register it is written to.