Trait oak_stage0::hal::Platform

pub trait Platform {
    type Mmio<S: PageSize>: Mmio<S>;

    // Required methods
    fn cpuid(leaf: u32) -> CpuidResult;
    unsafe fn mmio<S: PageSize>(base_address: PhysAddr) -> Self::Mmio<S>;
    fn port_factory() -> PortFactory;
    fn early_initialize_platform();
    fn initialize_platform(e820_table: &[BootE820Entry]);
    fn deinit_platform();
    fn populate_zero_page(zero_page: &mut ZeroPage);
    fn get_attestation(
        report_data: [u8; 64]
    ) -> Result<AttestationReport, &'static str>;
    fn get_derived_key() -> Result<DerivedKey, &'static str>;
    fn change_page_state(page: Page<Size4KiB>, state: PageAssignment);
    fn revalidate_page(page: Page<Size4KiB>);
    fn page_table_mask(encryption_state: PageEncryption) -> u64;
    fn encrypted() -> u64;
    fn tee_platform() -> TeePlatform;
    unsafe fn read_msr(msr: u32) -> u64;
    unsafe fn write_msr(msr: u32, value: u64);

}

Required Associated Types

type Mmio<S: PageSize>: Mmio<S>

Required Methods

fn cpuid(leaf: u32) -> CpuidResult

Performs the CPUID instruction.

unsafe fn mmio<S: PageSize>(base_address: PhysAddr) -> Self::Mmio<S>

fn port_factory() -> PortFactory

fn early_initialize_platform()

Platform-specific early initialization.

This sets up the bare minimum to get things going; for example, under SEV-ES and above, we set up the GHCB here, but nothing more.

This gets executed very soon after stage0 starts and comes with many restrictions:

You do not have access to logging.

You do not have access to heap allocator (BOOT_ALLOCATOR will still work).

fn initialize_platform(e820_table: &[BootE820Entry])

Platform-specific intialization.

This gets executed after early_initalize_platform() and some other auxiliary services, such as logging, have been set up; the main purpose is to accept all guest memory so that we can set up a heap allocator.

This does mean you do not have access to the heap allocator (BOOT_ALLOCATOR will still work).

fn deinit_platform()

Platform-specific cleanups just before stage0 jumps to the kernel.

The assumption is that after this operation there will no longer be any memory allocations or uses of the logging interface.

fn populate_zero_page(zero_page: &mut ZeroPage)

Populates platform-specific information in the zero page.

Example: locations of the SEV Secrets and CPUID pages.

fn get_attestation( report_data: [u8; 64] ) -> Result<AttestationReport, &'static str>

Returns an attestation report.

If AMD SEV-SNP is enabled it returns a valid hardware-rooted attestation report. In other cases it generates an empty attestation report for testing. The additional data will be set in both cases to bind the DICE chain to the attestation report.

Arguments

report_data - The custom data that must be included in the report. This is typically used to bind information (such as the hash of a public key) to the report.

fn get_derived_key() -> Result<DerivedKey, &'static str>

Requests a derived key.

The key is derived from the VCEK. The key derivation mixes in the VM launch measurement and guest policy and uses VMPL0.

We use this key as the unique device secret for deriving compound devices identifiers for each layer, and eventually a sealing key in the last layer.

fn change_page_state(page: Page<Size4KiB>, state: PageAssignment)

Ask for the page state to be changed by the hypervisor.

fn revalidate_page(page: Page<Size4KiB>)

Validate one page of memory.

This operation is required for SEV after going from a SHARED state to a PRIVATE state.

fn page_table_mask(encryption_state: PageEncryption) -> u64

Mask to use in the page tables for the given encrypion state.

SEV and TDX have opposite behaviours: for SEV, encrypted pages are marked; for TDX, unencrypted pages are marked.

fn encrypted() -> u64

Encrypted/shared bit mask irrespective of its semantics.

fn tee_platform() -> TeePlatform

unsafe fn read_msr(msr: u32) -> u64

Read a MSR.

Safety

The caller must guarantee that the MSR is valid.

unsafe fn write_msr(msr: u32, value: u64)

Write to a MSR.

Safety

The caller must guarantee that the MSR is valid.

Object Safety

This trait is not object safe.