Module oak_stage0::msr

Structs

ApicBase

The APIC Base Address Register.

ApicBaseFlags

Flags in the APIC Base Address Register (MSR 0x1B)

ApicErrorFlags

Flags in the APIC Error Status Register (offset 0x280)

MTRRDefType

IA32_MTRR_DefType base model specific register. See https://wiki.osdev.org/MTRR for documentation.

MTRRDefTypeFlags

Flags of the MTRRDefType Register.

SpuriousInterruptFlags

Flags in the Spurious Interrupt Register (offset 0x0F0)

X2ApicErrorStatusRegister

x2APIC Error Status Register.

X2ApicIdRegister

The x2APIC_ID register.

X2ApicInterruptCommandRegister

x2APIC Interrupt Command Register.

X2ApicSpuriousInterruptRegister

x2APIC Spurious Interrupt Register

X2ApicVersionRegister

x2APIC APIC Version Register

Enums

DestinationMode

Values for the destination mode flag in the Interrupt Command Register.

DestinationShorthand

Values for the destination shorthand flag in the Interrupt Command Register.

Level

Values for the level flag in the Interrupt Command Register.

MemoryType

The cache memory type used with MTRR. We only use Write-Protect mode which the Linux kernel expects to be enabled by the firmware in order to enable SEV.

MessageType

Interrupt types that can be sent via the Interrupt Command Register.

TriggerMode

Values for the trigger mode flag in the Interrupt Command Register.