Module oak_stage0::paging

Structs

PageTable

Thin wrapper around x86::PageTable that uses our PageTableEntry type.

PageTableEntry

Thin wrapper around x86_64::PageTableEntry that forces use of PageEncryption for addresses.

PageTableRefs

Wrapper for the page table references so that we can access them via a mutex rather than directly via unsafe code.

Enums

PageEncryption

Encryption state of a page in the page table.

Statics

PAGE_TABLE_REFS

References to all the pages tables we care about.

PDPT

PD_0

PD_3

PML4

Functions

init_page_table_refs

Initialises the page table references.

map_additional_memory

Maps the first 1GiB of memory using 2MiB hugepages, except for the first 2MiB that was already mapped as 512 4KiB pages.

remap_first_huge_page